+++ docs/config.json
@@ -15,6 +15,7 @@
                 "FeatureBuildingOptions",
                 "Style",
                 "StyleOptions",
+                "StyleContext",
                 "Label"
             ],
